+++ src/gamspy/_symbols/implicits/implicit_parameter.py
@@ -96,6 +96,9 @@ class ImplicitParameter(ImplicitSymbol, operable.Operable):
     def __ne__(self, other):  # type: ignore
         return expression.Expression(self, "ne", other)
 
+    def __repr__(self) -> str:
+        return f"ImplicitParameter(parent={self.parent}, name={self.name}, domain={self.domain}, permutation={self.permutation}), parent_scalar_domains={self.parent_scalar_domains})"
+
     @property
     def T(self) -> ImplicitParameter:
         """See gamspy.ImplicitParameter.t"""