+++ src/gamspy/_symbols/universe_alias.py
@@ -104,3 +104,6 @@ class UniverseAlias(gt.UniverseAlias):
             return len(self.records.index)
 
         return 0
+
+    def __repr__(self) -> str:
+        return f"UniverseAlias(container={repr(self.container)}, name={self.name})"